feat: modularizing to build each example in a separated way
