+++ callouts/rust/extproc/examples/add-header/main.rs
@@ -1,35 +1,24 @@
-// src/processor.rs
 use async_trait::async_trait;
-use crate::envoy::service::ext_proc::v3::{ProcessingRequest, ProcessingResponse};
-use thiserror::Error;
+use ext_proc::{
+    processor::{ExtProcessor, ProcessingError},
+    envoy::service::ext_proc::v3::{ProcessingRequest, ProcessingResponse},
+    service::ExtProcService,
+    utils::mutations,
+};
+use log::info;
 
-use crate::utils::mutations;
-
-#[derive(Error, Debug)]
-pub enum ProcessingError {
-    #[error("Processing failed: {0}")]
-    Failed(String),
-}
-
-#[async_trait]
-pub trait ExtProcessor: Send + Sync + 'static {
-    async fn process_request_headers(&self, req: &ProcessingRequest) -> Result<ProcessingResponse, ProcessingError>;
-    async fn process_response_headers(&self, req: &ProcessingRequest) -> Result<ProcessingResponse, ProcessingError>;
-    async fn process_request_body(&self, req: &ProcessingRequest) -> Result<ProcessingResponse, ProcessingError>;
-    async fn process_response_body(&self, req: &ProcessingRequest) -> Result<ProcessingResponse, ProcessingError>;
-}
-
+/// AddHeaderProcessor adds custom headers to both requests and responses
 #[derive(Clone)]
-pub struct HeaderMutationProcessor;
+struct AddHeaderProcessor;
 
-impl HeaderMutationProcessor {
-    pub fn new() -> Self {
+impl AddHeaderProcessor {
+    fn new() -> Self {
         Self
     }
 }
 
 #[async_trait]
-impl ExtProcessor for HeaderMutationProcessor {
+impl ExtProcessor for AddHeaderProcessor {
     async fn process_request_headers(&self, _req: &ProcessingRequest) -> Result<ProcessingResponse, ProcessingError> {
         Ok(mutations::add_header_mutation(
             vec![("header-request", "Value-request")],
@@ -53,4 +42,18 @@ impl ExtProcessor for HeaderMutationProcessor {
     async fn process_response_body(&self, _req: &ProcessingRequest) -> Result<ProcessingResponse, ProcessingError> {
         Ok(ProcessingResponse::default())
     }
+}
+
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
+    env_logger::init();
+
+    let addr = "0.0.0.0:8080";
+    let processor = AddHeaderProcessor::new();
+    let service = ExtProcService::new(processor);
+
+    info!("Starting add-header ext_proc service on {}", addr);
+    service.run(addr).await?;
+
+    Ok(())
 }
