+++ callouts/rust/extproc/src/utils/mutations.rs
@@ -1,4 +1,3 @@
-// src/utils/mutations.rs
 use crate::envoy::{
     config::core::v3::{HeaderValue, HeaderValueOption},
     service::ext_proc::v3::{
