--- callouts/rust/src/utils/mod.rs
@@ -1,2 +1,0 @@
-// src/utils/mod.rs
-pub mod mutations;
